Move queryable files to appropriate folders
