+++ HousingSearchApi/V1/Gateways/Models/Persons/QueryablePersonTenure.cs
@@ -1,6 +1,6 @@
 using Nest;
 
-namespace HousingSearchApi.V1.Gateways.Models
+namespace HousingSearchApi.V1.Gateways.Models.Persons
 {
     public class QueryablePersonTenure
     {
